requirements.txt
openai>=1.0.0
pandas>=1.3.0 
fi-instrumentation-otel==0.1.3